pyproject.toml
[tool.poetry]
name = "data-transfer"
version = "0.1.0"
description = ""
authors = ["Thomas Stocks <tstocks@mercedesamgf1.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
faust-streaming = "^0.10.5"
aiomonitor = "^0.4.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
